Changed int() cast to float().

+++ payCalculator.py
@@ -6,7 +6,7 @@ def calculatePay():
     # print("calculating pay")
     # This first line is provided for you
     hrs = int(input("Enter Hours: "))
-    rate = int(input("Enter Rate: "))
+    rate = float(input("Enter Rate: "))
     gross_pay = hrs * rate
     print(gross_pay)
